+++ R/rdb_providers.R
@@ -35,6 +35,7 @@
 #' rdb_providers(curl_config = list(proxy = "<proxy>", proxyport = <port>))
 #' }
 #' @seealso \code{\link{rdb_last_updates}}
+#' @author Sebastien Galais
 #' @export
 rdb_providers <- function(
   code = FALSE, use_readLines = getOption("rdbnomics.use_readLines"),